Fix compiler warning in the test suite from amd64 system
size_t is not an int. Cast size_t to unsigned long.
Also changed print type of %i to %lu.

Reviewed-on: https://gerrit.stanford.edu/135
Tested-by: Russ Allbery <rra@stanford.edu>

+++ tests/lib/base64-t.c
@@ -40,7 +40,8 @@ main(void)
         s = webauth_base64_encode(orig_buffer, i,
                                   encoded_buffer, &elen, BUFSIZE);
         rlen = webauth_base64_encoded_length(i);
-        is_int(WA_ERR_NONE, s, "Encoding length %i succeeds", i);
+        is_int(WA_ERR_NONE, s, "Encoding length %lu succeeds",
+                (unsigned long) i);
         is_int(rlen, elen, "...and returns the correct length");
 
         s = webauth_base64_decoded_length(encoded_buffer, elen, &rdlen);
@@ -48,7 +49,8 @@ main(void)
 
         s = webauth_base64_decode(encoded_buffer, elen, 
                                   decoded_buffer, &dlen, BUFSIZE);
-        is_int(WA_ERR_NONE, s, "Decoding length %i succeeds", i);
+        is_int(WA_ERR_NONE, s, "Decoding length %lu succeeds",
+                (unsigned long) i);
         is_int(rdlen, dlen, "...and returns the right length");
         is_int(i, dlen, "...which matches the original size");
         ok(memcmp(decoded_buffer, orig_buffer, i) == 0, "...and data");
